refactor: model field logic moved to a separate module.
