+++ pydantic_xml/__init__.py
@@ -4,8 +4,9 @@ pydantic xml serialization/deserialization extension
 
 from . import config, errors, model
 from .errors import ModelError, ParsingError
-from .model import BaseXmlModel, RootXmlModel, XmlFieldSerializer, XmlFieldValidator, attr, computed_attr
-from .model import computed_element, create_model, element, wrapped, xml_field_serializer, xml_field_validator
+from .fields import XmlFieldSerializer, XmlFieldValidator, attr, computed_attr, computed_element, element, wrapped
+from .fields import xml_field_serializer, xml_field_validator
+from .model import BaseXmlModel, RootXmlModel, create_model
 
 __all__ = (
     'BaseXmlModel',
